refactor-entertainment: update import statement for categoryRoutes
