+++ entertainment/src/api/routes/category-routes.ts
